refactoring Use cases structure
