+++ rover/usecases/src/main/kotlin/com/marsrover/nasa/rover/application/GetAllRovers.kt
@@ -1,0 +1,8 @@
+package com.marsrover.nasa.rover.application
+
+import com.marsrover.nasa.rover.domain.Rover
+
+interface GetAllRovers {
+    fun execute(): List<Rover>?
+}
+
