--- rover/usecases/src/main/kotlin/com/marsrover/nasa/rover/application/RoverCommandExecutor.kt
@@ -1,8 +1,0 @@
-package com.marsrover.nasa.rover.application
-
-import com.marsrover.nasa.rover.domain.RoverId
-
-class RoverCommandExecutor(roverId: RoverId, command: RoverCommandsExecService.Command) {
-
-
-}
