--- rover/usecases/src/main/kotlin/com/marsrover/nasa/rover/application/port/in/ExecuteRoversCommandsUseCase.kt
@@ -1,11 +1,0 @@
-package com.marsrover.nasa.rover.application.port.`in`
-
-import com.marsrover.nasa.rover.application.RoverCommandsExecService
-import com.marsrover.nasa.rover.domain.RoverId
-
-interface ExecuteRoversCommandsUseCase {
-    fun execute( roverId: RoverId, command: RoverCommandsExecService.Command) {
-
-    }
-}
-
