+++ rover/usecases/src/main/kotlin/com/marsrover/nasa/rover/application/scenaries/CommandUndoUseCase.kt
@@ -1,8 +1,8 @@
-package com.marsrover.nasa.rover.application
+package com.marsrover.nasa.rover.application.scenaries
 
 import java.util.*
 
-class CommandUndoService {
+class CommandUndoUseCase {
     private val commandStack: Stack<RoverCommandsExecService.Command> = Stack()
     fun addCommandToStack(command: RoverCommandsExecService.Command) {
         commandStack.push(command)
